disk: move partition creation to disk package
